Refactor server selection into its own trait.
