+++ app/Commands/Concerns/InteractsWithIO.php
@@ -117,33 +117,6 @@ trait InteractsWithIO
         );
     }
 
-    public function selectServer(string $action): Collection
-    {
-        $serverId = $this->argument('server_id');
-
-        if (empty($serverId)) {
-            $serverId = $this->askToSelectServer("Which server would you like to $action");
-        }
-
-        $server = $this->spinupwp->getServer((int) $serverId);
-
-        if ($this->forceOrConfirm("Are you sure you want to $action \"{$server->name}\"?")) {
-            return collect([$server]);
-        }
-
-        return collect();
-    }
-
-    public function askToSelectServer(string $question): int
-    {
-        $choices = collect($this->spinupwp->listServers());
-
-        return $this->askToSelect(
-            $question,
-            $choices->keyBy('id')->map(fn ($server) => $server->name)->toArray()
-        );
-    }
-
     /**
      * @param mixed $default
      */
